Add support to switch to the Nether or The End, and select properly
Import of models exported from the Nether and The End using Import
Settings now switch to these world types on import. Only models exported
from this or later versions will contain the "View Nether" and "View The
End" information needed for this to work.

Automatic height adjustment when using Hide obscured (usually on in the
Nether) now takes this feature into account, to give a better bottom
starting height.